Move const_drop_in_place test from ptr to manually_drop module
