+++ library/coretests/tests/ptr.rs
@@ -1045,41 +1045,3 @@ fn test_ptr_default() {
     let default = PtrMutDefaultTest::default();
     assert!(default.ptr.is_null());
 }
-
-#[test]
-fn test_const_drop_in_place() {
-    const COUNTER: usize = {
-        use core::cell::Cell;
-
-        let counter = Cell::new(0);
-
-        // only exists to make `Drop` indirect impl
-        #[allow(dead_code)]
-        struct Test<'a>(Dropped<'a>);
-
-        struct Dropped<'a>(&'a Cell<usize>);
-        impl const Drop for Dropped<'_> {
-            fn drop(&mut self) {
-                self.0.set(self.0.get() + 1);
-            }
-        }
-
-        let mut one = ManuallyDrop::new(Test(Dropped(&counter)));
-        let mut two = ManuallyDrop::new(Test(Dropped(&counter)));
-        let mut three = ManuallyDrop::new(Test(Dropped(&counter)));
-        assert!(counter.get() == 0);
-        unsafe {
-            ManuallyDrop::drop(&mut one);
-        }
-        assert!(counter.get() == 1);
-        unsafe {
-            ManuallyDrop::drop(&mut two);
-        }
-        assert!(counter.get() == 2);
-        unsafe {
-            ManuallyDrop::drop(&mut three);
-        }
-        counter.get()
-    };
-    assert_eq!(COUNTER, 3);
-}
